sr: demodevice: Make read channel non-blocking
Both pipe channels are currently configured as blocking. We read from the pipe
in receive_data. Since the channel is configured as blocking we'll block in
receive_data until all data has been received. receive_data will be called from
the mainloop, so as consequence the mainloop will be blocked until the demo
device has finished sampling. This is not so much of a if we are sampling in
blocking mode (using sr_session_run()) and the demo device is the only device in
the session, but it will fail badly for all other configurations (e.g. multiple
devices or async sampling).

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>