sr: session: Add session source
Currently the only method of running a session is to call sr_session_run().
sr_session_run() will block internally, which means it wont return until the
session has finished sampling. This is in particular annoying for GUI
applications, because the whole GUI will be come inaccessible during this time
and it is for example not possible to stop the session half way through.

This patch implements a GSource for a sigrok session. This allows to attach and
execute the session in a GMainContext, which means that it can be run
asynchronously.

The session source uses the same timeout scheme which is used for
sr_session_run(). This scheme is unfortunately broken if sources with different
timeouts exist because only the source with the lowest timeout will only ever
timeout. Fixing this will require a API rework though, so this is something that
needs to be done in subsequent patches for both blocking and non-blocking
sampling mode.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>