gtk: Add support for asynchronous sampling
Add support for asynchronous sampling to the application by attaching the
session to the applications mainloop. This keeps the gui accessible while
sampling is in progress and for example allows to abort a transfer.

The "sample" button is changed into a toggle button and will be kept active
during sampling. If the button is deactivated will sampling is still in
progress it will be aborted.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>